backend/requirements.txt
Flask>=2.0
flask-cors>=3.0
celery>=5.0
redis>=4.0
pytesseract>=0.3
PyMuPDF>=1.18.0
pdfplumber>=0.5.0
Pillow>=8.0
torch>=1.10
transformers>=4.0
gunicorn
redis
accelerate
